move BandReader to bands.reader
