+++ lib/python/bands/__init__.py
@@ -1,172 +1,1 @@
-import os
-import sys
-import json
-import glob
-from collections import OrderedDict
-
-import grass.script as gs
-
-class BandReaderError(Exception):
-    pass
-
-class BandReader:
-    """Band references reader"""
-    def __init__(self):
-        self._json_files = glob.glob(
-            os.path.join(os.environ['GISBASE'], 'etc', 'g.bands', '*.json')
-        )
-        if not self._json_files:
-            raise BandReaderError("No band definitions found")
-
-        self._read_config()
-
-    def _read_config(self):
-        """Read configuration"""
-        self.config = dict()
-        for json_file in self._json_files:
-            try:
-                with open(json_file) as fd:
-                    config = json.load(
-                        fd,
-                        object_pairs_hook=OrderedDict
-                    )
-            except json.decoder.JSONDecodeError as e:
-                raise BandReaderError(
-                    "Unable to parse '{}': {}".format(
-                        json_file, e
-                ))
-
-            # check if configuration is valid
-            self._check_config(config)
-
-            self.config[os.path.basename(json_file)] = config
-
-    @staticmethod
-    def _check_config(config):
-        """Check if config is valid
-
-        :todo: check shortcut uniqueness
-
-        :param dict config: configuration to be validated
-        """
-        for items in config.values():
-            for item in ('shortcut', 'bands'):
-                if item not in items.keys():
-                    raise BandReaderError(
-                        "Invalid band definition: <{}> is missing".format(item
-                ))
-            if len(items['bands']) < 1:
-                raise BandReaderError(
-                    "Invalid band definition: no bands defined"
-                )
-
-    @staticmethod
-    def _print_band_extended(band, item):
-        """Print band-specific metadata
-
-        :param str band: band identifier
-        :param str item: items to be printed out
-        """
-        indent = 4
-        print ('{}band: {}'.format(
-            ' ' * indent, band
-        ))
-        for k, v in item[band].items():
-            print ('{}{}: {}'.format(' ' * indent * 2, k, v))
-
-    def _print_band(self, shortcut, band):
-        print (self._band_identifier(shortcut, band))
-
-    def print_info(self, shortcut=None, band=None, extended=False):
-        """Prints band reference information to stdout.
-
-        Can be filtered by shortcut or band identifier.
-
-        :param str shortcut: shortcut to filter (eg. S2A) or None
-        :param str band: band (eg. 1) or None
-        :param bool extended: print also extended metadata
-        """
-        found = False
-        for root in self.config.values():
-            for item in root.values():
-                if shortcut and \
-                   item['shortcut'][0:len(shortcut)].upper() != shortcut.upper():
-                    continue
-
-                found = True
-                if band and band not in item['bands']:
-                    raise BandReaderError(
-                        "Band <{}> not found in <{}>".format(
-                            band, shortcut
-                        ))
-
-                # print generic information
-                if extended:
-                    for subitem in item.keys():
-                        if subitem == 'bands':
-                            # bands item is processed bellow
-                            continue
-                        print ('{}: {}'.format(
-                            subitem, item[subitem]
-                        ))
-
-                    # print detailed band information
-                    if band:
-                        self._print_band_extended(band, item['bands'])
-                    else:
-                        for iband in item['bands']:
-                            self._print_band_extended(iband, item['bands'])
-                else:
-                    if band:
-                        self._print_band(item['shortcut'], band)
-                    else:
-                        for iband in item['bands']:
-                            self._print_band(item['shortcut'], iband)
-
-        # raise error when defined shortcut not found
-        if shortcut and not found:
-            raise BandReaderError(
-                "Band reference <{}> not found".format(shortcut)
-            )
-
-    def find_file(self, band_reference):
-        """Find file by band reference.
-
-        Match is case-insensitive.
-
-        :param str band_reference: band reference identifier to search for (eg. S2A_1)
-
-        :return str: file basename if found or None
-        """
-        try:
-            shortcut, band = band_reference.split('_')
-        except ValueError:
-            raise BandReaderError("Invalid band identifier <{}>".format(
-                band_reference
-            ))
-
-        for filename, config in self.config.items():
-            for root in config.keys():
-                if config[root]['shortcut'].upper() == shortcut.upper() and \
-                   band.upper() in map(lambda x: x.upper(), config[root]['bands'].keys()):
-                    return filename
-
-        return None
-
-    def get_bands(self):
-        """Get list of band identifiers.
-
-        :return list: list of valid band identifiers
-        """
-        bands = []
-        for root in self.config.values():
-            for item in root.values():
-                for band in item['bands']:
-                    bands.append(
-                        self._band_identifier(item['shortcut'], band)
-                    )
-        return bands
-
-    @staticmethod
-    def _band_identifier(shortcut, band):
-        return '{}_{}'.format(shortcut, band)
+from .reader import BandReader, BandReaderError
